+++ test/nn/encoder/test_stype_encoder.py
@@ -16,7 +16,6 @@ from torch_frame.nn import (
     LinearEmbeddingEncoder,
     LinearEncoder,
     LinearModelEncoder,
-    LinearMultiEmbeddingEncoder,
     LinearPeriodicEncoder,
     MultiCategoricalEmbeddingEncoder,
     StackEncoder,
@@ -255,8 +254,7 @@ def test_text_embedded_encoder():
         stypes=[
             torch_frame.text_embedded,
         ],
-        # TODO: Allow different text embedder for each column
-        text_embedder_cfg=TextEmbedderConfig(
+        col_to_text_embedder_cfg=TextEmbedderConfig(
             text_embedder=HashTextEmbedder(text_emb_channels),
             batch_size=None),
     )
@@ -266,7 +264,6 @@ def test_text_embedded_encoder():
         dataset.col_stats[col_name]
         for col_name in tensor_frame.col_names_dict[stype.text_embedded]
     ]
-    # TODO: Use LinearMultiEmbeddingEncoder instead
     encoder = LinearEmbeddingEncoder(
         out_channels=out_channels,
         stats_list=stats_list,
@@ -296,7 +293,7 @@ def test_embedding_encoder():
         dataset.col_stats[col_name]
         for col_name in tensor_frame.col_names_dict[stype.embedding]
     ]
-    encoder = LinearMultiEmbeddingEncoder(
+    encoder = LinearEmbeddingEncoder(
         out_channels=out_channels,
         stats_list=stats_list,
         stype=stype.embedding,