+++ test/nn/encoder/test_stypewise_encoder.py
@@ -11,7 +11,6 @@ from torch_frame.nn import (
     LinearEmbeddingEncoder,
     LinearEncoder,
     LinearModelEncoder,
-    LinearMultiEmbeddingEncoder,
     LinearPeriodicEncoder,
     MultiCategoricalEmbeddingEncoder,
     StypeWiseFeatureEncoder,
@@ -34,24 +33,18 @@ from torch_frame.testing.text_tokenizer import (
 @pytest.mark.parametrize('encoder_multicategorical_cls_kwargs', [
     (MultiCategoricalEmbeddingEncoder, {}),
 ])
-@pytest.mark.parametrize(
-    'encoder_text_embedded_cls_kwargs',
-    [
-        # TODO: Migrate to LinearMultiEmbeddingEncoder
-        (LinearEmbeddingEncoder, {}),
-    ])
+@pytest.mark.parametrize('encoder_text_embedded_cls_kwargs', [
+    (LinearEmbeddingEncoder, {}),
+])
 @pytest.mark.parametrize('encoder_text_tokenized_cls_kwargs', [
     (LinearModelEncoder, {
         'model': RandomTextModel(12, 2),
         'in_channels': 12,
     }),
 ])
-@pytest.mark.parametrize(
-    'encoder_embedding_cls_kwargs',
-    [
-        # TODO: Migrate to LinearMultiEmbeddingEncoder
-        (LinearMultiEmbeddingEncoder, {}),
-    ])
+@pytest.mark.parametrize('encoder_embedding_cls_kwargs', [
+    (LinearEmbeddingEncoder, {}),
+])
 def test_stypewise_feature_encoder(
     encoder_cat_cls_kwargs,
     encoder_num_cls_kwargs,
@@ -72,7 +65,7 @@ def test_stypewise_feature_encoder(
             stype.text_tokenized,
             stype.embedding,
         ],
-        text_embedder_cfg=TextEmbedderConfig(
+        col_to_text_embedder_cfg=TextEmbedderConfig(
             text_embedder=HashTextEmbedder(out_channels=16, ),
             batch_size=None,
         ),