+++ test/utils/test_io.py
@@ -23,7 +23,7 @@ def teardown_module():
 
 def get_fake_dataset(
     num_rows: int,
-    text_embedder_cfg: TextEmbedderConfig,
+    col_to_text_embedder_cfg: TextEmbedderConfig,
     text_tokenizer_cfg: TextTokenizerConfig,
 ) -> FakeDataset:
     stypes = [
@@ -38,7 +38,7 @@ def get_fake_dataset(
     dataset = FakeDataset(
         num_rows=num_rows,
         stypes=stypes,
-        text_embedder_cfg=text_embedder_cfg,
+        col_to_text_embedder_cfg=col_to_text_embedder_cfg,
         text_tokenizer_cfg=text_tokenizer_cfg,
     )
     return dataset
@@ -48,13 +48,13 @@ def test_dataset_cache():
     num_rows = 10
     out_channels = 8
 
-    text_embedder_cfg = TextEmbedderConfig(
+    col_to_text_embedder_cfg = TextEmbedderConfig(
         text_embedder=HashTextEmbedder(out_channels))
     text_tokenizer_cfg = TextTokenizerConfig(
         text_tokenizer=WhiteSpaceHashTokenizer())
     dataset = get_fake_dataset(
         num_rows,
-        text_embedder_cfg,
+        col_to_text_embedder_cfg,
         text_tokenizer_cfg,
     )
 
@@ -63,7 +63,7 @@ def test_dataset_cache():
 
     new_dataset = get_fake_dataset(
         num_rows,
-        text_embedder_cfg,
+        col_to_text_embedder_cfg,
         text_tokenizer_cfg,
     )
     new_dataset.df = dataset.df
@@ -83,7 +83,7 @@ def test_dataset_cache():
 
     new_dataset = get_fake_dataset(
         num_rows,
-        text_embedder_cfg,
+        col_to_text_embedder_cfg,
         text_tokenizer_cfg,
     )
     new_dataset.df = new_dataset.df
@@ -98,13 +98,14 @@ def test_dataset_cache():
 def test_save_load_tensor_frame():
     num_rows = 10
     out_channels = 8
-    text_embedder_cfg = TextEmbedderConfig(
+    col_to_text_embedder_cfg = TextEmbedderConfig(
         text_embedder=HashTextEmbedder(out_channels))
     text_tokenizer_cfg = TextTokenizerConfig(
         text_tokenizer=WhiteSpaceHashTokenizer(),
         batch_size=None,
     )
-    dataset = get_fake_dataset(num_rows, text_embedder_cfg, text_tokenizer_cfg)
+    dataset = get_fake_dataset(num_rows, col_to_text_embedder_cfg,
+                               text_tokenizer_cfg)
     dataset.materialize()
 
     path = osp.join(TEST_DIR.name, TEST_SAVE_LOAD_NAME)