+++ torch_frame/nn/encoder/__init__.py
@@ -10,7 +10,6 @@ from .stype_encoder import (
     LinearPeriodicEncoder,
     ExcelFormerEncoder,
     LinearEmbeddingEncoder,
-    LinearMultiEmbeddingEncoder,
     LinearModelEncoder,
     StackEncoder,
 )
@@ -26,7 +25,6 @@ __all__ = classes = [
     'LinearPeriodicEncoder',
     'ExcelFormerEncoder',
     'LinearEmbeddingEncoder',
-    'LinearMultiEmbeddingEncoder',
     'LinearModelEncoder',
     'StackEncoder',
 ]