+++ torch_frame/nn/encoder/stype_encoder.py
@@ -570,65 +570,9 @@ class ExcelFormerEncoder(StypeEncoder):
 
 
 class LinearEmbeddingEncoder(StypeEncoder):
-    r"""Linear function based encoder for pre-computed embedding features with
-    the same dimensionalities. It applies a linear layer
-    :obj:`torch.nn.Linear(emb_dim, out_channels)` on each embedding feature and
-    concatenates the output embeddings. Note that the implementation does this
-    for all :obj:`text_embedded` features in a batched manner.
-    NOTE: We will soon shift this to a new LinearMultiEmbeddingEncoder that
-    takes :class:`MultiEmbeddingTensor` object as input.
-    """
-    # TODO: To be deprecated in favaor of LinearMultiEmbeddingEncoder
-    supported_stypes = {stype.text_embedded}
-
-    def __init__(
-        self,
-        out_channels: Optional[int] = None,
-        stats_list: Optional[List[Dict[StatType, Any]]] = None,
-        stype: Optional[stype] = None,
-        post_module: Optional[Module] = None,
-        na_strategy: Optional[NAStrategy] = None,
-    ):
-        super().__init__(out_channels, stats_list, stype, post_module,
-                         na_strategy)
-
-    def init_modules(self):
-        super().init_modules()
-        num_cols = len(self.stats_list)
-        emb_dim_list = [stats[StatType.EMB_DIM] for stats in self.stats_list]
-        if not all(emb_dim == emb_dim_list[0] for emb_dim in emb_dim_list):
-            # TODO: Relax this by using feat: MultiEmbeddedTensor
-            raise RuntimeError(
-                "All embeddings must be of the same dimensionality.")
-        emb_dim = emb_dim_list[0]
-
-        self.weight = Parameter(
-            torch.empty(num_cols, emb_dim, self.out_channels))
-        self.bias = Parameter(torch.empty(num_cols, self.out_channels))
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        super().reset_parameters()
-        torch.nn.init.normal_(self.weight, std=0.01)
-        torch.nn.init.zeros_(self.bias)
-
-    def encode_forward(self, feat: Tensor) -> Tensor:
-        # [batch_size, num_cols, in_channels] *
-        # [num_cols, in_channels, out_channels]
-        # -> [batch_size, num_cols, out_channels]
-        x_lin = torch.einsum("ijk,jkl->ijl", feat, self.weight)
-        # [batch_size, num_cols, out_channels] + [num_cols, out_channels]
-        # -> [batch_size, num_cols, out_channels]
-        x = x_lin + self.bias
-        return x
-
-
-class LinearMultiEmbeddingEncoder(StypeEncoder):
     r"""Linear function based encoder for pre-computed embedding features.
     It applies a linear layer :obj:`torch.nn.Linear(emb_dim, out_channels)`
-    on each embedding feature and concatenates the output embeddings. Note that
-    the implementation does this for all :obj:`text_embedded` features in a
-    batched manner.
+    on each embedding feature and concatenates the output embeddings.
     """
     supported_stypes = {stype.text_embedded, stype.embedding}