+++ torch_frame/utils/io.py
@@ -42,12 +42,12 @@ def deserialize_feat_dict(
     for stype, feat_serialized in feat_serialized_dict.items():
         if stype.use_multi_nested_tensor:
             feat = MultiNestedTensor(**feat_serialized)
+        elif stype.use_multi_embedding_tensor:
+            feat = MultiEmbeddingTensor(**feat_serialized)
         elif stype.use_dict_multi_nested_tensor:
             feat = {}
             for name, f_serialized in feat_serialized.items():
                 feat[name] = MultiNestedTensor(**f_serialized)
-        elif stype.use_multi_embedding_tensor:
-            feat = MultiEmbeddingTensor(**feat_serialized)
         else:
             assert isinstance(feat_serialized, Tensor)
             feat = feat_serialized